+++ test/truffle/cexts/json/json.sh
@@ -2,9 +2,9 @@
 
 source test/truffle/common.sh.inc
 
-jt gem-test-pack
+gem_test_pack=$(jt gem-test-pack)
 
-jt ruby -S gem install --local "$(jt gem-test-pack)/gem-cache/json-2.2.0.gem" -V -N --backtrace
+jt ruby -S gem install --local "$gem_test_pack/gem-cache/json-2.2.0.gem" -V -N --backtrace
 
 output=$(jt --silent ruby -e 'gem "json"; require "json"; puts JSON.dump({ a: 1 })')