+++ test/truffle/cexts/unf_ext/unf_ext.sh
@@ -2,8 +2,8 @@
 
 source test/truffle/common.sh.inc
 
-jt gem-test-pack
+gem_test_pack=$(jt gem-test-pack)
 
-jt ruby -S gem install --local "$(jt gem-test-pack)/gem-cache/unf_ext-0.0.7.4.gem" -V -N --backtrace
+jt ruby -S gem install --local "$gem_test_pack/gem-cache/unf_ext-0.0.7.4.gem" -V -N --backtrace
 
 jt ruby test/truffle/cexts/unf_ext/test.rb